refactor: update model logo imports to use new modelLogos configuration
