+++ src/renderer/src/store/migrate.ts
@@ -1,13 +1,9 @@
 import { loggerService } from '@logger'
 import { nanoid } from '@reduxjs/toolkit'
 import { DEFAULT_CONTEXTCOUNT, DEFAULT_TEMPERATURE, isMac } from '@renderer/config/constant'
+import { glm45FlashModel, SYSTEM_MODELS } from '@renderer/config/defaultModel'
 import { DEFAULT_MIN_APPS } from '@renderer/config/minapps'
-import {
-  glm45FlashModel,
-  isFunctionCallingModel,
-  isNotSupportedTextDelta,
-  SYSTEM_MODELS
-} from '@renderer/config/models'
+import { isFunctionCallingModel, isNotSupportedTextDelta } from '@renderer/config/models'
 import { BUILTIN_OCR_PROVIDERS, BUILTIN_OCR_PROVIDERS_MAP, DEFAULT_OCR_PROVIDER } from '@renderer/config/ocr'
 import { TRANSLATE_PROMPT } from '@renderer/config/prompts'
 import {
@@ -2356,11 +2352,14 @@ const migrateConfig = {
       if (state.settings && state.note) {
         const showWorkspaceValue = (state.settings as any)?.showWorkspace
         if (showWorkspaceValue !== undefined) {
+          // @ts-ignore eslint-disable-next-line
           state.note.settings.showWorkspace = showWorkspaceValue
           // Remove from settings
           delete (state.settings as any).showWorkspace
+          // @ts-ignore eslint-disable-next-line
         } else if (state.note.settings.showWorkspace === undefined) {
           // Set default value if not exists
+          // @ts-ignore eslint-disable-next-line
           state.note.settings.showWorkspace = true
         }
       }
